fix: delete the api-connection files to simplify code
